requirements.txt
# Safe installation requirements
# Install these one by one if needed

# Essential packages first
pip>=23.0
setuptools>=65.0
wheel>=0.37.0

# Core data science (most stable versions)
pandas==1.5.3
numpy==1.24.3

# Machine learning
scikit-learn==1.2.2

# Visualization
matplotlib==3.7.1
plotly==5.14.1

# Streamlit
streamlit==1.28.1

# Additional utilities
pyyaml==6.0
openpyxl==3.1.2